+++ Reflect-App/source/reflect/addons/reflect-form-mailer/src/js/transpiled/settings 3.js
@@ -1,17 +1,17 @@
 "use strict";
 
 /**
- * Reflect
- * @package Reflect
+ * Reflect Form Mailer Addon
+ * @package ReflectFormMailerAddon
  * @author Min Tat Goy <david@davidgoy.dev>
- * @link https://github.com/davidgoy/reflect
+ * @link https://github.com/davidgoy/reflect-form-mailer
  * @copyright 2020 Min Tat Goy
  * @license https://www.gnu.org/licenses/gpl.html   GPLv2 or later
- * @version 1.0.0-beta.5
+ * @version 1.0.0-beta.4
  * @since File available since v1.0.0-alpha.1
  */
 window.addEventListener('DOMContentLoaded', function () {
-  (function reflectSetup() {
+  (function reflectFormMailerAddonSettings() {
     var csrfPreventionToken = document.querySelector('#csrfPreventionToken').dataset.csrfPreventionToken; //--------------------------------------------------------------------------
 
     /**
@@ -19,28 +19,21 @@ window.addEventListener('DOMContentLoaded', function () {
      */
     //--------------------------------------------------------------------------
 
-    (function handleSiteKeyRegeneration() {
-      var regenerateSiteKeyButton = document.querySelector('#regenerateSiteKey');
-      regenerateSiteKeyButton.addEventListener('click', function (event) {
-        var formData = new FormData();
-        formData.append('doAsync', 'generateSiteKey');
-        formData.append('csrfPreventionToken', csrfPreventionToken); // Debug
+    (function initSwitches() {
+      var switches = document.getElementsByClassName('switch');
 
-        /*
-        for(let input of formData.entries()) {
-           console.log('Input name: ' + input[0] + ' | Input value: ' + input[1]);
-        }
-        */
+      for (var i = 0; i < switches.length; i++) {
+        switches[i].addEventListener('change', function (event) {
+          var theSwitch = event.target;
+          var theActualInput = document.querySelector("input[name=".concat(theSwitch.id, "]")); // This is a hidden input. We do it this way because Bootstrap currently uses checkbox input type to render a switch
 
-        apiGetData(formData).then(function (response) {
-          if (response !== 'false') {
-            var siteKeyInput = document.querySelector('#siteKey');
-            siteKeyInput.value = response;
-          } else {// Debug
-            //console.log(`Error: ${response}`);
+          if (theSwitch.checked === true) {
+            theActualInput.value = 'true';
+          } else {
+            theActualInput.value = 'false';
           }
         });
-      });
+      }
     })(); //--------------------------------------------------------------------------
 
     /**
@@ -49,11 +42,12 @@ window.addEventListener('DOMContentLoaded', function () {
     //--------------------------------------------------------------------------
 
 
-    (function handleSetupFormSubmission() {
-      var form = document.querySelector('#setup');
+    (function handleSettingsFormSubmission() {
+      var form = document.querySelector('#reflectFormMailerAddonSettings');
       form.addEventListener('submit', function (event) {
         var formData = new FormData(form);
-        formData.append('doAsync', 'saveSiteKey');
+        formData.append('doAsync', 'saveAddonSettings');
+        formData.append('addonFolderName', 'reflect-form-mailer');
         formData.append('csrfPreventionToken', csrfPreventionToken); // Debug
 
         /*
@@ -65,35 +59,19 @@ window.addEventListener('DOMContentLoaded', function () {
         apiGetData(formData).then(function (response) {
           if (response === 'true') {
             Swal.fire({
-              title: 'SITE KEY SAVED',
-              text: 'Have you copied and stored the site key somewhere safe?',
+              title: 'SETTINGS SAVED',
               icon: 'success',
-              showCancelButton: true,
-              confirmButtonText: 'YES',
-              cancelButtonText: 'NO',
-              reverseButtons: true
-            }).then(function (userConfirmation) {
-              if (userConfirmation.value) {
-                // Reload the page
-                location.reload();
-              } else if (userConfirmation.dismiss === Swal.DismissReason.cancel) {// Do nothing
-              }
+              confirmButtonText: 'OK'
             });
           } else {
             Swal.fire({
-              title: 'ERROR',
-              text: 'Site key may have already been saved. Otherwise check config.json for possible errors.',
+              title: 'SETTINGS NOT SAVED',
+              text: 'Either site key cannot be verified or there is a problem with config.json.',
               icon: 'error',
               confirmButtonText: 'OK'
-            }).then(function (userConfirmation) {
-              if (userConfirmation.value) {
-                // Reload the page
-                location.reload();
-              }
             });
           }
-        }); // Prevent the form from submitting in the usual way (which would trigger a http request)
-
+        });
         event.preventDefault();
       });
     })(); //--------------------------------------------------------------------------