+++ Reflect-App/source/reflect/core/src/js/transpiled/setup 3.js
@@ -1,17 +1,17 @@
 "use strict";
 
 /**
- * Reflect Form Mailer Addon
- * @package ReflectFormMailerAddon
+ * Reflect
+ * @package Reflect
  * @author Min Tat Goy <david@davidgoy.dev>
- * @link https://github.com/davidgoy/reflect-form-mailer
+ * @link https://github.com/davidgoy/reflect
  * @copyright 2020 Min Tat Goy
  * @license https://www.gnu.org/licenses/gpl.html   GPLv2 or later
- * @version 1.0.0-beta.5
+ * @version 1.0.0-beta.4
  * @since File available since v1.0.0-alpha.1
  */
 window.addEventListener('DOMContentLoaded', function () {
-  (function reflectFormMailerAddonSettings() {
+  (function reflectSetup() {
     var csrfPreventionToken = document.querySelector('#csrfPreventionToken').dataset.csrfPreventionToken; //--------------------------------------------------------------------------
 
     /**
@@ -19,21 +19,28 @@ window.addEventListener('DOMContentLoaded', function () {
      */
     //--------------------------------------------------------------------------
 
-    (function initSwitches() {
-      var switches = document.getElementsByClassName('switch');
+    (function handleSiteKeyRegeneration() {
+      var regenerateSiteKeyButton = document.querySelector('#regenerateSiteKey');
+      regenerateSiteKeyButton.addEventListener('click', function (event) {
+        var formData = new FormData();
+        formData.append('doAsync', 'generateSiteKey');
+        formData.append('csrfPreventionToken', csrfPreventionToken); // Debug
 
-      for (var i = 0; i < switches.length; i++) {
-        switches[i].addEventListener('change', function (event) {
-          var theSwitch = event.target;
-          var theActualInput = document.querySelector("input[name=".concat(theSwitch.id, "]")); // This is a hidden input. We do it this way because Bootstrap currently uses checkbox input type to render a switch
+        /*
+        for(let input of formData.entries()) {
+           console.log('Input name: ' + input[0] + ' | Input value: ' + input[1]);
+        }
+        */
 
-          if (theSwitch.checked === true) {
-            theActualInput.value = 'true';
-          } else {
-            theActualInput.value = 'false';
+        apiGetData(formData).then(function (response) {
+          if (response !== 'false') {
+            var siteKeyInput = document.querySelector('#siteKey');
+            siteKeyInput.value = response;
+          } else {// Debug
+            //console.log(`Error: ${response}`);
           }
         });
-      }
+      });
     })(); //--------------------------------------------------------------------------
 
     /**
@@ -42,12 +49,11 @@ window.addEventListener('DOMContentLoaded', function () {
     //--------------------------------------------------------------------------
 
 
-    (function handleSettingsFormSubmission() {
-      var form = document.querySelector('#reflectFormMailerAddonSettings');
+    (function handleSetupFormSubmission() {
+      var form = document.querySelector('#setup');
       form.addEventListener('submit', function (event) {
         var formData = new FormData(form);
-        formData.append('doAsync', 'saveAddonSettings');
-        formData.append('addonFolderName', 'reflect-form-mailer');
+        formData.append('doAsync', 'saveSiteKey');
         formData.append('csrfPreventionToken', csrfPreventionToken); // Debug
 
         /*
@@ -59,19 +65,35 @@ window.addEventListener('DOMContentLoaded', function () {
         apiGetData(formData).then(function (response) {
           if (response === 'true') {
             Swal.fire({
-              title: 'SETTINGS SAVED',
+              title: 'SITE KEY SAVED',
+              text: 'Have you copied and stored the site key somewhere safe?',
               icon: 'success',
-              confirmButtonText: 'OK'
+              showCancelButton: true,
+              confirmButtonText: 'YES',
+              cancelButtonText: 'NO',
+              reverseButtons: true
+            }).then(function (userConfirmation) {
+              if (userConfirmation.value) {
+                // Reload the page
+                location.reload();
+              } else if (userConfirmation.dismiss === Swal.DismissReason.cancel) {// Do nothing
+              }
             });
           } else {
             Swal.fire({
-              title: 'SETTINGS NOT SAVED',
-              text: 'Either site key cannot be verified or there is a problem with config.json.',
+              title: 'ERROR',
+              text: 'Site key may have already been saved. Otherwise check config.json for possible errors.',
               icon: 'error',
               confirmButtonText: 'OK'
+            }).then(function (userConfirmation) {
+              if (userConfirmation.value) {
+                // Reload the page
+                location.reload();
+              }
             });
           }
-        });
+        }); // Prevent the form from submitting in the usual way (which would trigger a http request)
+
         event.preventDefault();
       });
     })(); //--------------------------------------------------------------------------