+++ Reflect-App/source/reflect/core/src/js/transpiled/static-files-manager 3.js
@@ -9,7 +9,7 @@ function _readOnlyError(name) { throw new Error("\"" + name + "\" is read-only")
  * @link https://github.com/davidgoy/reflect
  * @copyright 2020 Min Tat Goy
  * @license https://www.gnu.org/licenses/gpl.html   GPLv2 or later
- * @version 1.0.0-beta.6
+ * @version 1.0.0-beta.4
  * @since File available since v1.0.0-alpha.1
  */
 window.addEventListener('DOMContentLoaded', function () {